Added modular configurations. Added back ishare.
